lara+stats: handle distance travelled overflows
This ensures distance travelled is always treated as an unsigned
value, so in extreme cases it will wrap safely back to 0. When Lara is
performing certain glitches that put her Y position out of bounds,
the stats will continue to use her previous Y position to avoid
skewing the results.

Resolves TRX1368.

+++ src/trx/game/stats/common.c
@@ -265,7 +265,8 @@ void Stats_AddDistanceTravelled(const XYZ_32 pos, const XYZ_32 last_pos)
 {
     LEVEL_STATS *const stats = Stats_GetLevelStats(Game_GetCurrentLevel());
     if (stats != nullptr) {
-        stats->distance_travelled += XYZ_32_GetDistance(pos, last_pos);
+        stats->distance_travelled +=
+            (uint32_t)XYZ_32_GetDistance(pos, last_pos);
     }
 }
